# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WildGuard proposed backend API",
    "version": "1.0.0",
    "description": "Original proposed full-system contract. Only milestone 1 is implemented; use implemented-openapi.json for its current envelope and APIs. Other modules remain future work."
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Planned full-system API; see implemented-openapi.json for the implemented subset."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/media/{id}": {
      "put": {
        "tags": [
          "Media"
        ],
        "summary": "Upload an image with a stable client UUID",
        "operationId": "put_Media_12",
        "x-roles": [
          "PARK_MANAGER",
          "RANGER",
          "LIAISON_OFFICER",
          "RESEARCHER",
          "COMMUNITY_MEMBER"
        ],
        "description": "Multipart form fields file, parkId and category. Decode/validate JPEG or PNG, max 5 MiB. Category must match role: ranger INCIDENT/ALERT_RESPONSE; liaison ALERT_RESPONSE; community member COMMUNITY_REPORT; researcher/manager CAMERA_TRAP. Same UUID, owner, category, park and bytes returns 200; changed bytes/metadata is 409. Only owner can access unlinked uploads.",
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                },
                "example": {
                  "id": "33333333-3333-4333-8333-333333333333",
                  "parkId": "park-yala",
                  "category": "INCIDENT",
                  "contentType": "image/jpeg",
                  "sizeBytes": 245800,
                  "contentUrl": "/api/v1/media/33333333-3333-4333-8333-333333333333/content",
                  "createdAt": "2026-10-08T06:41:00Z"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "URI of the created resource.",
                "schema": {
                  "type": "string"
                },
                "example": "/api/v1/media/33333333-3333-4333-8333-333333333333"
              }
            }
          },
          "200": {
            "description": "Identical replay returns the existing record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                },
                "example": {
                  "id": "33333333-3333-4333-8333-333333333333",
                  "parkId": "park-yala",
                  "category": "INCIDENT",
                  "contentType": "image/jpeg",
                  "sizeBytes": 245800,
                  "contentUrl": "/api/v1/media/33333333-3333-4333-8333-333333333333/content",
                  "createdAt": "2026-10-08T06:41:00Z"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "415": {
            "$ref": "#/components/responses/Error415"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "33333333-3333-4333-8333-333333333333"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/MediaUpload"
              },
              "example": {
                "file": "<JPEG binary bytes>",
                "parkId": "park-yala",
                "category": "INCIDENT"
              }
            }
          }
        }
      }
    },
    "/media/{id}/content": {
      "get": {
        "tags": [
          "Media"
        ],
        "summary": "Download authorized image content",
        "operationId": "get_Media_13",
        "x-roles": [
          "PARK_MANAGER",
          "RANGER",
          "LIAISON_OFFICER",
          "RESEARCHER",
          "COMMUNITY_MEMBER"
        ],
        "description": "Owner or caller authorized to read the linked parent record. Unlinked media are owner-only. Authenticated download; never expose local file paths.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                },
                "example": "<JPEG or PNG bytes>"
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "33333333-3333-4333-8333-333333333333"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "timestamp",
          "status",
          "code",
          "message",
          "path"
        ],
        "additionalProperties": false
      },
...
      "MediaUpload": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "description": "JPEG or PNG image, at most 5 MiB; decoded image validation required."
          },
          "parkId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": ".*\\S.*"
          },
          "category": {
            "type": "string",
            "enum": [
              "INCIDENT",
              "COMMUNITY_REPORT",
              "CAMERA_TRAP",
              "ALERT_RESPONSE"
            ]
          }
        },
        "required": [
          "file",
          "parkId",
          "category"
        ],
        "additionalProperties": false
      },
      "Media": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parkId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": ".*\\S.*"
          },
          "category": {
            "type": "string",
            "enum": [
              "INCIDENT",
              "COMMUNITY_REPORT",
              "CAMERA_TRAP",
              "ALERT_RESPONSE"
            ]
          },
          "contentType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png"
            ]
          },
          "sizeBytes": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5242880
          },
          "contentUrl": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "parkId",
          "category",
          "contentType",
          "sizeBytes",
          "contentUrl",
          "createdAt"
        ],
        "additionalProperties": false
      },
...
    },
    "responses": {
      "Error400": {
        "description": "Invalid request or validation failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 400,
              "code": "VALIDATION_FAILED",
              "message": "Invalid request or validation failed.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error401": {
        "description": "Missing/invalid token, inactive account or invalid login.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 401,
              "code": "UNAUTHORIZED",
              "message": "Missing/invalid token, inactive account or invalid login.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error403": {
        "description": "Role or requested park is forbidden.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 403,
              "code": "FORBIDDEN",
              "message": "Role or requested park is forbidden.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error404": {
        "description": "Resource not found or not accessible.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 404,
              "code": "NOT_FOUND",
              "message": "Resource not found or not accessible.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error409": {
        "description": "Conflicting state or resource ID reused with different data.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 409,
              "code": "STATE_CONFLICT",
              "message": "Conflicting state or resource ID reused with different data.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error413": {
        "description": "Image exceeds 5 MiB upload limit.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 413,
              "code": "FILE_TOO_LARGE",
              "message": "Image exceeds 5 MiB upload limit.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error415": {
        "description": "Unsupported or invalid image content.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 415,
              "code": "UNSUPPORTED_MEDIA_TYPE",
              "message": "Unsupported or invalid image content.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error500": {
        "description": "Unexpected server error; no false success acknowledgement.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 500,
              "code": "INTERNAL_ERROR",
              "message": "Unexpected server error; no false success acknowledgement.",
              "path": "/api/v1/example"
            }
          }
        }
      },
      "Error503": {
        "description": "Database/storage temporarily unavailable; retry later.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "timestamp": "2026-10-08T06:41:00Z",
              "status": 503,
              "code": "SERVICE_UNAVAILABLE",
              "message": "Database/storage temporarily unavailable; retry later.",
              "path": "/api/v1/example"
            }
          }
        }
      }
    }
  }
}